+++ pkg/ingester/limiter_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 
+	"github.com/cortexproject/cortex/pkg/util"
 	"github.com/cortexproject/cortex/pkg/util/validation"
 )
 
@@ -17,65 +18,124 @@ func TestSeriesLimit_maxSeriesPerMetric(t *testing.T) {
 		maxLocalSeriesPerMetric  int
 		maxGlobalSeriesPerMetric int
 		ringReplicationFactor    int
+		ringZoneAwarenessEnabled bool
 		ringIngesterCount        int
+		ringZonesCount           int
 		shardByAllLabels         bool
-		expected                 int
+		shardSize                int
+		expectedDefaultSharding  int
+		expectedShuffleSharding  int
 	}{
 		"both local and global limits are disabled": {
 			maxLocalSeriesPerMetric:  0,
 			maxGlobalSeriesPerMetric: 0,
 			ringReplicationFactor:    1,
 			ringIngesterCount:        1,
+			ringZonesCount:           1,
 			shardByAllLabels:         false,
-			expected:                 math.MaxInt32,
+			expectedDefaultSharding:  math.MaxInt32,
+			expectedShuffleSharding:  math.MaxInt32,
 		},
 		"only local limit is enabled": {
 			maxLocalSeriesPerMetric:  1000,
 			maxGlobalSeriesPerMetric: 0,
 			ringReplicationFactor:    1,
 			ringIngesterCount:        1,
+			ringZonesCount:           1,
 			shardByAllLabels:         false,
-			expected:                 1000,
+			expectedDefaultSharding:  1000,
+			expectedShuffleSharding:  1000,
 		},
 		"only global limit is enabled with shard-by-all-labels=false and replication-factor=1": {
 			maxLocalSeriesPerMetric:  0,
 			maxGlobalSeriesPerMetric: 1000,
 			ringReplicationFactor:    1,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         false,
-			expected:                 1000,
+			shardSize:                5,
+			expectedDefaultSharding:  1000,
+			expectedShuffleSharding:  1000,
 		},
 		"only global limit is enabled with shard-by-all-labels=true and replication-factor=1": {
 			maxLocalSeriesPerMetric:  0,
 			maxGlobalSeriesPerMetric: 1000,
 			ringReplicationFactor:    1,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         true,
-			expected:                 100,
+			shardSize:                5,
+			expectedDefaultSharding:  100,
+			expectedShuffleSharding:  200,
 		},
 		"only global limit is enabled with shard-by-all-labels=true and replication-factor=3": {
 			maxLocalSeriesPerMetric:  0,
 			maxGlobalSeriesPerMetric: 1000,
 			ringReplicationFactor:    3,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         true,
-			expected:                 300,
+			shardSize:                5,
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  600,
 		},
 		"both local and global limits are set with local limit < global limit": {
 			maxLocalSeriesPerMetric:  150,
 			maxGlobalSeriesPerMetric: 1000,
 			ringReplicationFactor:    3,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         true,
-			expected:                 150,
+			shardSize:                5,
+			expectedDefaultSharding:  150,
+			expectedShuffleSharding:  150,
 		},
 		"both local and global limits are set with local limit > global limit": {
-			maxLocalSeriesPerMetric:  500,
+			maxLocalSeriesPerMetric:  800,
 			maxGlobalSeriesPerMetric: 1000,
 			ringReplicationFactor:    3,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         true,
-			expected:                 300,
+			shardSize:                5,
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  600,
+		},
+		"zone-awareness enabled, global limit enabled and the shard size is NOT divisible by number of zones": {
+			maxLocalSeriesPerMetric:  0,
+			maxGlobalSeriesPerMetric: 900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                5, // Not divisible by number of zones.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  450, // (900 / 6) * 3
+		},
+		"zone-awareness enabled, global limit enabled and the shard size is divisible by number of zones": {
+			maxLocalSeriesPerMetric:  0,
+			maxGlobalSeriesPerMetric: 900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                6, // Divisible by number of zones.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  450, // (900 / 6) * 3
+		},
+		"zone-awareness enabled, global limit enabled and the shard size > number of ingesters": {
+			maxLocalSeriesPerMetric:  0,
+			maxGlobalSeriesPerMetric: 900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                20, // Greater than number of ingesters.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  300,
 		},
 	}
 
@@ -86,18 +146,25 @@ func TestSeriesLimit_maxSeriesPerMetric(t *testing.T) {
 			// Mock the ring
 			ring := &ringCountMock{}
 			ring.On("HealthyInstancesCount").Return(testData.ringIngesterCount)
+			ring.On("ZonesCount").Return(testData.ringZonesCount)
 
 			// Mock limits
 			limits, err := validation.NewOverrides(validation.Limits{
 				MaxLocalSeriesPerMetric:  testData.maxLocalSeriesPerMetric,
 				MaxGlobalSeriesPerMetric: testData.maxGlobalSeriesPerMetric,
+				IngestionTenantShardSize: testData.shardSize,
 			}, nil)
 			require.NoError(t, err)
 
-			limiter := NewLimiter(limits, ring, testData.ringReplicationFactor, testData.shardByAllLabels)
+			// Assert on default sharding strategy.
+			limiter := NewLimiter(limits, ring, util.ShardingStrategyDefault, testData.shardByAllLabels, testData.ringReplicationFactor, testData.ringZoneAwarenessEnabled)
 			actual := limiter.maxSeriesPerMetric("test")
+			assert.Equal(t, testData.expectedDefaultSharding, actual)
 
-			assert.Equal(t, testData.expected, actual)
+			// Assert on shuffle sharding strategy.
+			limiter = NewLimiter(limits, ring, util.ShardingStrategyShuffle, testData.shardByAllLabels, testData.ringReplicationFactor, testData.ringZoneAwarenessEnabled)
+			actual = limiter.maxSeriesPerMetric("test")
+			assert.Equal(t, testData.expectedShuffleSharding, actual)
 		})
 	}
 }
@@ -107,65 +174,124 @@ func TestSeriesLimit_maxMetadataPerMetric(t *testing.T) {
 		maxLocalMetadataPerMetric  int
 		maxGlobalMetadataPerMetric int
 		ringReplicationFactor      int
+		ringZoneAwarenessEnabled   bool
 		ringIngesterCount          int
+		ringZonesCount             int
 		shardByAllLabels           bool
-		expected                   int
+		shardSize                  int
+		expectedDefaultSharding    int
+		expectedShuffleSharding    int
 	}{
 		"both local and global limits are disabled": {
 			maxLocalMetadataPerMetric:  0,
 			maxGlobalMetadataPerMetric: 0,
 			ringReplicationFactor:      1,
 			ringIngesterCount:          1,
+			ringZonesCount:             1,
 			shardByAllLabels:           false,
-			expected:                   math.MaxInt32,
+			expectedDefaultSharding:    math.MaxInt32,
+			expectedShuffleSharding:    math.MaxInt32,
 		},
 		"only local limit is enabled": {
 			maxLocalMetadataPerMetric:  1000,
 			maxGlobalMetadataPerMetric: 0,
 			ringReplicationFactor:      1,
 			ringIngesterCount:          1,
+			ringZonesCount:             1,
 			shardByAllLabels:           false,
-			expected:                   1000,
+			expectedDefaultSharding:    1000,
+			expectedShuffleSharding:    1000,
 		},
 		"only global limit is enabled with shard-by-all-labels=false and replication-factor=1": {
 			maxLocalMetadataPerMetric:  0,
 			maxGlobalMetadataPerMetric: 1000,
 			ringReplicationFactor:      1,
 			ringIngesterCount:          10,
+			ringZonesCount:             1,
 			shardByAllLabels:           false,
-			expected:                   1000,
+			shardSize:                  5,
+			expectedDefaultSharding:    1000,
+			expectedShuffleSharding:    1000,
 		},
 		"only global limit is enabled with shard-by-all-labels=true and replication-factor=1": {
 			maxLocalMetadataPerMetric:  0,
 			maxGlobalMetadataPerMetric: 1000,
 			ringReplicationFactor:      1,
 			ringIngesterCount:          10,
+			ringZonesCount:             1,
 			shardByAllLabels:           true,
-			expected:                   100,
+			shardSize:                  5,
+			expectedDefaultSharding:    100,
+			expectedShuffleSharding:    200,
 		},
 		"only global limit is enabled with shard-by-all-labels=true and replication-factor=3": {
 			maxLocalMetadataPerMetric:  0,
 			maxGlobalMetadataPerMetric: 1000,
 			ringReplicationFactor:      3,
 			ringIngesterCount:          10,
+			ringZonesCount:             1,
 			shardByAllLabels:           true,
-			expected:                   300,
+			shardSize:                  5,
+			expectedDefaultSharding:    300,
+			expectedShuffleSharding:    600,
 		},
 		"both local and global limits are set with local limit < global limit": {
 			maxLocalMetadataPerMetric:  150,
 			maxGlobalMetadataPerMetric: 1000,
 			ringReplicationFactor:      3,
 			ringIngesterCount:          10,
+			ringZonesCount:             1,
 			shardByAllLabels:           true,
-			expected:                   150,
+			shardSize:                  5,
+			expectedDefaultSharding:    150,
+			expectedShuffleSharding:    150,
 		},
 		"both local and global limits are set with local limit > global limit": {
-			maxLocalMetadataPerMetric:  500,
+			maxLocalMetadataPerMetric:  800,
 			maxGlobalMetadataPerMetric: 1000,
 			ringReplicationFactor:      3,
 			ringIngesterCount:          10,
+			ringZonesCount:             1,
 			shardByAllLabels:           true,
-			expected:                   300,
+			shardSize:                  5,
+			expectedDefaultSharding:    300,
+			expectedShuffleSharding:    600,
+		},
+		"zone-awareness enabled, global limit enabled and the shard size is NOT divisible by number of zones": {
+			maxLocalMetadataPerMetric:  0,
+			maxGlobalMetadataPerMetric: 900,
+			ringReplicationFactor:      3,
+			ringZoneAwarenessEnabled:   true,
+			ringIngesterCount:          9,
+			ringZonesCount:             3,
+			shardByAllLabels:           true,
+			shardSize:                  5, // Not divisible by number of zones.
+			expectedDefaultSharding:    300,
+			expectedShuffleSharding:    450, // (900 / 6) * 3
+		},
+		"zone-awareness enabled, global limit enabled and the shard size is divisible by number of zones": {
+			maxLocalMetadataPerMetric:  0,
+			maxGlobalMetadataPerMetric: 900,
+			ringReplicationFactor:      3,
+			ringZoneAwarenessEnabled:   true,
+			ringIngesterCount:          9,
+			ringZonesCount:             3,
+			shardByAllLabels:           true,
+			shardSize:                  6, // Divisible by number of zones.
+			expectedDefaultSharding:    300,
+			expectedShuffleSharding:    450, // (900 / 6) * 3
+		},
+		"zone-awareness enabled, global limit enabled and the shard size > number of ingesters": {
+			maxLocalMetadataPerMetric:  0,
+			maxGlobalMetadataPerMetric: 900,
+			ringReplicationFactor:      3,
+			ringZoneAwarenessEnabled:   true,
+			ringIngesterCount:          9,
+			ringZonesCount:             3,
+			shardByAllLabels:           true,
+			shardSize:                  20, // Greater than number of ingesters.
+			expectedDefaultSharding:    300,
+			expectedShuffleSharding:    300,
 		},
 	}
 
@@ -176,86 +302,151 @@ func TestSeriesLimit_maxMetadataPerMetric(t *testing.T) {
 			// Mock the ring
 			ring := &ringCountMock{}
 			ring.On("HealthyInstancesCount").Return(testData.ringIngesterCount)
+			ring.On("ZonesCount").Return(testData.ringZonesCount)
 
 			// Mock limits
 			limits, err := validation.NewOverrides(validation.Limits{
 				MaxLocalMetadataPerMetric:  testData.maxLocalMetadataPerMetric,
 				MaxGlobalMetadataPerMetric: testData.maxGlobalMetadataPerMetric,
+				IngestionTenantShardSize:   testData.shardSize,
 			}, nil)
 			require.NoError(t, err)
 
-			limiter := NewLimiter(limits, ring, testData.ringReplicationFactor, testData.shardByAllLabels)
+			// Assert on default sharding strategy.
+			limiter := NewLimiter(limits, ring, util.ShardingStrategyDefault, testData.shardByAllLabels, testData.ringReplicationFactor, testData.ringZoneAwarenessEnabled)
 			actual := limiter.maxMetadataPerMetric("test")
+			assert.Equal(t, testData.expectedDefaultSharding, actual)
 
-			assert.Equal(t, testData.expected, actual)
+			// Assert on default shuffle strategy.
+			limiter = NewLimiter(limits, ring, util.ShardingStrategyShuffle, testData.shardByAllLabels, testData.ringReplicationFactor, testData.ringZoneAwarenessEnabled)
+			actual = limiter.maxMetadataPerMetric("test")
+			assert.Equal(t, testData.expectedShuffleSharding, actual)
 		})
 	}
 }
 
 func TestSeriesLimit_maxSeriesPerUser(t *testing.T) {
 	tests := map[string]struct {
-		maxLocalSeriesPerUser  int
-		maxGlobalSeriesPerUser int
-		ringReplicationFactor  int
-		ringIngesterCount      int
-		shardByAllLabels       bool
-		expected               int
+		maxLocalSeriesPerUser    int
+		maxGlobalSeriesPerUser   int
+		ringReplicationFactor    int
+		ringZoneAwarenessEnabled bool
+		ringIngesterCount        int
+		ringZonesCount           int
+		shardByAllLabels         bool
+		shardSize                int
+		expectedDefaultSharding  int
+		expectedShuffleSharding  int
 	}{
 		"both local and global limits are disabled": {
-			maxLocalSeriesPerUser:  0,
-			maxGlobalSeriesPerUser: 0,
-			ringReplicationFactor:  1,
-			ringIngesterCount:      1,
-			shardByAllLabels:       false,
-			expected:               math.MaxInt32,
+			maxLocalSeriesPerUser:   0,
+			maxGlobalSeriesPerUser:  0,
+			ringReplicationFactor:   1,
+			ringIngesterCount:       1,
+			ringZonesCount:          1,
+			shardByAllLabels:        false,
+			expectedDefaultSharding: math.MaxInt32,
+			expectedShuffleSharding: math.MaxInt32,
 		},
 		"only local limit is enabled": {
-			maxLocalSeriesPerUser:  1000,
-			maxGlobalSeriesPerUser: 0,
-			ringReplicationFactor:  1,
-			ringIngesterCount:      1,
-			shardByAllLabels:       false,
-			expected:               1000,
+			maxLocalSeriesPerUser:   1000,
+			maxGlobalSeriesPerUser:  0,
+			ringReplicationFactor:   1,
+			ringIngesterCount:       1,
+			ringZonesCount:          1,
+			shardByAllLabels:        false,
+			expectedDefaultSharding: 1000,
+			expectedShuffleSharding: 1000,
 		},
 		"only global limit is enabled with shard-by-all-labels=false and replication-factor=1": {
-			maxLocalSeriesPerUser:  0,
-			maxGlobalSeriesPerUser: 1000,
-			ringReplicationFactor:  1,
-			ringIngesterCount:      10,
-			shardByAllLabels:       false,
-			expected:               math.MaxInt32,
+			maxLocalSeriesPerUser:   0,
+			maxGlobalSeriesPerUser:  1000,
+			ringReplicationFactor:   1,
+			ringIngesterCount:       10,
+			ringZonesCount:          1,
+			shardByAllLabels:        false,
+			expectedDefaultSharding: math.MaxInt32,
+			expectedShuffleSharding: math.MaxInt32,
 		},
 		"only global limit is enabled with shard-by-all-labels=true and replication-factor=1": {
-			maxLocalSeriesPerUser:  0,
-			maxGlobalSeriesPerUser: 1000,
-			ringReplicationFactor:  1,
-			ringIngesterCount:      10,
-			shardByAllLabels:       true,
-			expected:               100,
+			maxLocalSeriesPerUser:   0,
+			maxGlobalSeriesPerUser:  1000,
+			ringReplicationFactor:   1,
+			ringIngesterCount:       10,
+			ringZonesCount:          1,
+			shardByAllLabels:        true,
+			shardSize:               5,
+			expectedDefaultSharding: 100,
+			expectedShuffleSharding: 200,
 		},
 		"only global limit is enabled with shard-by-all-labels=true and replication-factor=3": {
-			maxLocalSeriesPerUser:  0,
-			maxGlobalSeriesPerUser: 1000,
-			ringReplicationFactor:  3,
-			ringIngesterCount:      10,
-			shardByAllLabels:       true,
-			expected:               300,
+			maxLocalSeriesPerUser:   0,
+			maxGlobalSeriesPerUser:  1000,
+			ringReplicationFactor:   3,
+			ringIngesterCount:       10,
+			ringZonesCount:          1,
+			shardByAllLabels:        true,
+			shardSize:               5,
+			expectedDefaultSharding: 300,
+			expectedShuffleSharding: 600,
 		},
 		"both local and global limits are set with local limit < global limit": {
-			maxLocalSeriesPerUser:  150,
-			maxGlobalSeriesPerUser: 1000,
-			ringReplicationFactor:  3,
-			ringIngesterCount:      10,
-			shardByAllLabels:       true,
-			expected:               150,
+			maxLocalSeriesPerUser:   150,
+			maxGlobalSeriesPerUser:  1000,
+			ringReplicationFactor:   3,
+			ringIngesterCount:       10,
+			ringZonesCount:          1,
+			shardByAllLabels:        true,
+			shardSize:               5,
+			expectedDefaultSharding: 150,
+			expectedShuffleSharding: 150,
 		},
 		"both local and global limits are set with local limit > global limit": {
-			maxLocalSeriesPerUser:  500,
-			maxGlobalSeriesPerUser: 1000,
-			ringReplicationFactor:  3,
-			ringIngesterCount:      10,
-			shardByAllLabels:       true,
-			expected:               300,
+			maxLocalSeriesPerUser:   800,
+			maxGlobalSeriesPerUser:  1000,
+			ringReplicationFactor:   3,
+			ringIngesterCount:       10,
+			ringZonesCount:          1,
+			shardByAllLabels:        true,
+			shardSize:               5,
+			expectedDefaultSharding: 300,
+			expectedShuffleSharding: 600,
+		},
+		"zone-awareness enabled, global limit enabled and the shard size is NOT divisible by number of zones": {
+			maxLocalSeriesPerUser:    0,
+			maxGlobalSeriesPerUser:   900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                5, // Not divisible by number of zones.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  450, // (900 / 6) * 3
+		},
+		"zone-awareness enabled, global limit enabled and the shard size is divisible by number of zones": {
+			maxLocalSeriesPerUser:    0,
+			maxGlobalSeriesPerUser:   900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                6, // Divisible by number of zones.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  450, // (900 / 6) * 3
+		},
+		"zone-awareness enabled, global limit enabled and the shard size > number of ingesters": {
+			maxLocalSeriesPerUser:    0,
+			maxGlobalSeriesPerUser:   900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                20, // Greater than number of ingesters.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  300,
 		},
 	}
 
@@ -266,18 +457,25 @@ func TestSeriesLimit_maxSeriesPerUser(t *testing.T) {
 			// Mock the ring
 			ring := &ringCountMock{}
 			ring.On("HealthyInstancesCount").Return(testData.ringIngesterCount)
+			ring.On("ZonesCount").Return(testData.ringZonesCount)
 
 			// Mock limits
 			limits, err := validation.NewOverrides(validation.Limits{
-				MaxLocalSeriesPerUser:  testData.maxLocalSeriesPerUser,
-				MaxGlobalSeriesPerUser: testData.maxGlobalSeriesPerUser,
+				MaxLocalSeriesPerUser:    testData.maxLocalSeriesPerUser,
+				MaxGlobalSeriesPerUser:   testData.maxGlobalSeriesPerUser,
+				IngestionTenantShardSize: testData.shardSize,
 			}, nil)
 			require.NoError(t, err)
 
-			limiter := NewLimiter(limits, ring, testData.ringReplicationFactor, testData.shardByAllLabels)
+			// Assert on default sharding strategy.
+			limiter := NewLimiter(limits, ring, util.ShardingStrategyDefault, testData.shardByAllLabels, testData.ringReplicationFactor, testData.ringZoneAwarenessEnabled)
 			actual := limiter.maxSeriesPerUser("test")
+			assert.Equal(t, testData.expectedDefaultSharding, actual)
 
-			assert.Equal(t, testData.expected, actual)
+			// Assert on shuffle sharding strategy.
+			limiter = NewLimiter(limits, ring, util.ShardingStrategyShuffle, testData.shardByAllLabels, testData.ringReplicationFactor, testData.ringZoneAwarenessEnabled)
+			actual = limiter.maxSeriesPerUser("test")
+			assert.Equal(t, testData.expectedShuffleSharding, actual)
 		})
 	}
 }
@@ -287,65 +485,123 @@ func TestLimit_maxMetadataPerUser(t *testing.T) {
 		maxLocalMetadataPerUser  int
 		maxGlobalMetadataPerUser int
 		ringReplicationFactor    int
+		ringZoneAwarenessEnabled bool
 		ringIngesterCount        int
+		ringZonesCount           int
 		shardByAllLabels         bool
-		expected                 int
+		shardSize                int
+		expectedDefaultSharding  int
+		expectedShuffleSharding  int
 	}{
 		"both local and global limits are disabled": {
 			maxLocalMetadataPerUser:  0,
 			maxGlobalMetadataPerUser: 0,
 			ringReplicationFactor:    1,
 			ringIngesterCount:        1,
+			ringZonesCount:           1,
 			shardByAllLabels:         false,
-			expected:                 math.MaxInt32,
+			expectedDefaultSharding:  math.MaxInt32,
+			expectedShuffleSharding:  math.MaxInt32,
 		},
 		"only local limit is enabled": {
 			maxLocalMetadataPerUser:  1000,
 			maxGlobalMetadataPerUser: 0,
 			ringReplicationFactor:    1,
 			ringIngesterCount:        1,
+			ringZonesCount:           1,
 			shardByAllLabels:         false,
-			expected:                 1000,
+			expectedDefaultSharding:  1000,
+			expectedShuffleSharding:  1000,
 		},
 		"only global limit is enabled with shard-by-all-labels=false and replication-factor=1": {
 			maxLocalMetadataPerUser:  0,
 			maxGlobalMetadataPerUser: 1000,
 			ringReplicationFactor:    1,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         false,
-			expected:                 math.MaxInt32,
+			expectedDefaultSharding:  math.MaxInt32,
+			expectedShuffleSharding:  math.MaxInt32,
 		},
 		"only global limit is enabled with shard-by-all-labels=true and replication-factor=1": {
 			maxLocalMetadataPerUser:  0,
 			maxGlobalMetadataPerUser: 1000,
 			ringReplicationFactor:    1,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         true,
-			expected:                 100,
+			shardSize:                5,
+			expectedDefaultSharding:  100,
+			expectedShuffleSharding:  200,
 		},
 		"only global limit is enabled with shard-by-all-labels=true and replication-factor=3": {
 			maxLocalMetadataPerUser:  0,
 			maxGlobalMetadataPerUser: 1000,
 			ringReplicationFactor:    3,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         true,
-			expected:                 300,
+			shardSize:                5,
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  600,
 		},
 		"both local and global limits are set with local limit < global limit": {
 			maxLocalMetadataPerUser:  150,
 			maxGlobalMetadataPerUser: 1000,
 			ringReplicationFactor:    3,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         true,
-			expected:                 150,
+			shardSize:                5,
+			expectedDefaultSharding:  150,
+			expectedShuffleSharding:  150,
 		},
 		"both local and global limits are set with local limit > global limit": {
-			maxLocalMetadataPerUser:  500,
+			maxLocalMetadataPerUser:  800,
 			maxGlobalMetadataPerUser: 1000,
 			ringReplicationFactor:    3,
 			ringIngesterCount:        10,
+			ringZonesCount:           1,
 			shardByAllLabels:         true,
-			expected:                 300,
+			shardSize:                5,
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  600,
+		},
+		"zone-awareness enabled, global limit enabled and the shard size is NOT divisible by number of zones": {
+			maxLocalMetadataPerUser:  0,
+			maxGlobalMetadataPerUser: 900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                5, // Not divisible by number of zones.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  450, // (900 / 6) * 3
+		},
+		"zone-awareness enabled, global limit enabled and the shard size is divisible by number of zones": {
+			maxLocalMetadataPerUser:  0,
+			maxGlobalMetadataPerUser: 900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                6, // Divisible by number of zones.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  450, // (900 / 6) * 3
+		},
+		"zone-awareness enabled, global limit enabled and the shard size > number of ingesters": {
+			maxLocalMetadataPerUser:  0,
+			maxGlobalMetadataPerUser: 900,
+			ringReplicationFactor:    3,
+			ringZoneAwarenessEnabled: true,
+			ringIngesterCount:        9,
+			ringZonesCount:           3,
+			shardByAllLabels:         true,
+			shardSize:                20, // Greater than number of ingesters.
+			expectedDefaultSharding:  300,
+			expectedShuffleSharding:  300,
 		},
 	}
 
@@ -356,18 +612,25 @@ func TestLimit_maxMetadataPerUser(t *testing.T) {
 			// Mock the ring
 			ring := &ringCountMock{}
 			ring.On("HealthyInstancesCount").Return(testData.ringIngesterCount)
+			ring.On("ZonesCount").Return(testData.ringZonesCount)
 
 			// Mock limits
 			limits, err := validation.NewOverrides(validation.Limits{
 				MaxLocalMetricsWithMetadataPerUser:  testData.maxLocalMetadataPerUser,
 				MaxGlobalMetricsWithMetadataPerUser: testData.maxGlobalMetadataPerUser,
+				IngestionTenantShardSize:            testData.shardSize,
 			}, nil)
 			require.NoError(t, err)
 
-			limiter := NewLimiter(limits, ring, testData.ringReplicationFactor, testData.shardByAllLabels)
+			// Assert on default sharding strategy.
+			limiter := NewLimiter(limits, ring, util.ShardingStrategyDefault, testData.shardByAllLabels, testData.ringReplicationFactor, testData.ringZoneAwarenessEnabled)
 			actual := limiter.maxMetadataPerUser("test")
+			assert.Equal(t, testData.expectedDefaultSharding, actual)
 
-			assert.Equal(t, testData.expected, actual)
+			// Assert on shuffle sharding strategy.
+			limiter = NewLimiter(limits, ring, util.ShardingStrategyShuffle, testData.shardByAllLabels, testData.ringReplicationFactor, testData.ringZoneAwarenessEnabled)
+			actual = limiter.maxMetadataPerUser("test")
+			assert.Equal(t, testData.expectedShuffleSharding, actual)
 		})
 	}
 }
@@ -418,6 +681,7 @@ func TestLimiter_AssertMaxSeriesPerMetric(t *testing.T) {
 			// Mock the ring
 			ring := &ringCountMock{}
 			ring.On("HealthyInstancesCount").Return(testData.ringIngesterCount)
+			ring.On("ZonesCount").Return(1)
 
 			// Mock limits
 			limits, err := validation.NewOverrides(validation.Limits{
@@ -426,7 +690,7 @@ func TestLimiter_AssertMaxSeriesPerMetric(t *testing.T) {
 			}, nil)
 			require.NoError(t, err)
 
-			limiter := NewLimiter(limits, ring, testData.ringReplicationFactor, testData.shardByAllLabels)
+			limiter := NewLimiter(limits, ring, util.ShardingStrategyDefault, testData.shardByAllLabels, testData.ringReplicationFactor, false)
 			actual := limiter.AssertMaxSeriesPerMetric("test", testData.series)
 
 			assert.Equal(t, testData.expected, actual)
@@ -479,6 +743,7 @@ func TestLimiter_AssertMaxMetadataPerMetric(t *testing.T) {
 			// Mock the ring
 			ring := &ringCountMock{}
 			ring.On("HealthyInstancesCount").Return(testData.ringIngesterCount)
+			ring.On("ZonesCount").Return(1)
 
 			// Mock limits
 			limits, err := validation.NewOverrides(validation.Limits{
@@ -487,7 +752,7 @@ func TestLimiter_AssertMaxMetadataPerMetric(t *testing.T) {
 			}, nil)
 			require.NoError(t, err)
 
-			limiter := NewLimiter(limits, ring, testData.ringReplicationFactor, testData.shardByAllLabels)
+			limiter := NewLimiter(limits, ring, util.ShardingStrategyDefault, testData.shardByAllLabels, testData.ringReplicationFactor, false)
 			actual := limiter.AssertMaxMetadataPerMetric("test", testData.metadata)
 
 			assert.Equal(t, testData.expected, actual)
@@ -541,6 +806,7 @@ func TestLimiter_AssertMaxSeriesPerUser(t *testing.T) {
 			// Mock the ring
 			ring := &ringCountMock{}
 			ring.On("HealthyInstancesCount").Return(testData.ringIngesterCount)
+			ring.On("ZonesCount").Return(1)
 
 			// Mock limits
 			limits, err := validation.NewOverrides(validation.Limits{
@@ -549,7 +815,7 @@ func TestLimiter_AssertMaxSeriesPerUser(t *testing.T) {
 			}, nil)
 			require.NoError(t, err)
 
-			limiter := NewLimiter(limits, ring, testData.ringReplicationFactor, testData.shardByAllLabels)
+			limiter := NewLimiter(limits, ring, util.ShardingStrategyDefault, testData.shardByAllLabels, testData.ringReplicationFactor, false)
 			actual := limiter.AssertMaxSeriesPerUser("test", testData.series)
 
 			assert.Equal(t, testData.expected, actual)
@@ -603,6 +869,7 @@ func TestLimiter_AssertMaxMetricsWithMetadataPerUser(t *testing.T) {
 			// Mock the ring
 			ring := &ringCountMock{}
 			ring.On("HealthyInstancesCount").Return(testData.ringIngesterCount)
+			ring.On("ZonesCount").Return(1)
 
 			// Mock limits
 			limits, err := validation.NewOverrides(validation.Limits{
@@ -611,7 +878,7 @@ func TestLimiter_AssertMaxMetricsWithMetadataPerUser(t *testing.T) {
 			}, nil)
 			require.NoError(t, err)
 
-			limiter := NewLimiter(limits, ring, testData.ringReplicationFactor, testData.shardByAllLabels)
+			limiter := NewLimiter(limits, ring, util.ShardingStrategyDefault, testData.shardByAllLabels, testData.ringReplicationFactor, false)
 			actual := limiter.AssertMaxMetricsWithMetadataPerUser("test", testData.metadata)
 
 			assert.Equal(t, testData.expected, actual)
@@ -668,6 +935,11 @@ type ringCountMock struct {
 }
 
 func (m *ringCountMock) HealthyInstancesCount() int {
+	args := m.Called()
+	return args.Int(0)
+}
+
+func (m *ringCountMock) ZonesCount() int {
 	args := m.Called()
 	return args.Int(0)
 }