+++ pkg/ring/replication_set_test.go
@@ -1,0 +1,35 @@
+package ring
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestReplicationSet_GetAddresses(t *testing.T) {
+	tests := map[string]struct {
+		rs       ReplicationSet
+		expected []string
+	}{
+		"should return an empty slice on empty replication set": {
+			rs:       ReplicationSet{},
+			expected: []string{},
+		},
+		"should return instances addresses (no order guaranteed)": {
+			rs: ReplicationSet{
+				Ingesters: []IngesterDesc{
+					{Addr: "127.0.0.1"},
+					{Addr: "127.0.0.2"},
+					{Addr: "127.0.0.3"},
+				},
+			},
+			expected: []string{"127.0.0.1", "127.0.0.2", "127.0.0.3"},
+		},
+	}
+
+	for testName, testData := range tests {
+		t.Run(testName, func(t *testing.T) {
+			assert.ElementsMatch(t, testData.expected, testData.rs.GetAddresses())
+		})
+	}
+}